Compiler portability changes to files in alignment folder (so it will all compile, if not link, under MSVC): 1. Using boost::scoped_array (instead of stack-based     dynamic arrays), in MSVC builds.  In     Alignment::checkSeqName(),     Alignment::computeConst(), Alignment::countStates().     MSVC doesn't allow dynamic-sized on-stack arrays. 2. Changing the types of numerous variables to avoid     warnings about risky implicit casts.  Particularly in     Alignment::checkSeqName(). 3. Changing size_t variables to intptr_t, so that they can be     used in #pragma omp parallel for... in MSVC builds     (MSVC insists that the loop variables be signed).     e.g. in AlignmentSummary::constructSequenceMatrix. 4. Used local variables (local_df and local_ddf) in     AlignmentPairwise::computeFuncDerv (because the MS      omp implenmentation doesn't like you to use reference     variables as reduction targets!).

+++ alignment/alignmentsummary.cpp
@@ -13,7 +13,7 @@ AlignmentSummary::AlignmentSummary(const Alignment* a
                                    , bool keepBoringSites) {
     alignment      = a;
     sequenceMatrix = nullptr;
-    sequenceCount  = a->getNSeq();
+    sequenceCount  = static_cast<int>(a->getNSeq());
     totalFrequency = 0;
     totalFrequencyOfNonConstSites = 0;
     if (sequenceCount==0) {
@@ -32,20 +32,20 @@ AlignmentSummary::AlignmentSummary(const Alignment* a
         SiteSummary(): isConst(false), frequency(0), minState(0), maxState(0) {}
     };
     
-    size_t siteCount = alignment->size();
+    intptr_t siteCount = alignment->size();
     std::vector<SiteSummary> sites;
     sites.resize(siteCount);
     #ifdef _OPENMP
         #pragma omp parallel for
     #endif
-    for (size_t site=0; site<siteCount; ++site) { //per site
+    for (intptr_t site=0; site<siteCount; ++site) { //per site
         auto        itSite = alignment->begin() + site;
         SiteSummary &s     = sites[site];
         StateType minStateForSite = (*itSite)[0];
         StateType maxStateForSite = minStateForSite;
         s.isConst   = itSite->isConst();
         s.frequency = itSite->frequency;
-        for (size_t seq=1; seq<sequenceCount; ++seq) {
+        for (int seq=1; seq<sequenceCount; ++seq) {
             auto state = (*itSite)[seq] ;
             if (state < minStateForSite ) {
                 minStateForSite = state;
@@ -59,7 +59,7 @@ AlignmentSummary::AlignmentSummary(const Alignment* a
     }
     sequenceLength = 0; //Number sites where there's some variability
     std::map<int, int> & map = stateToSumOfConstantSiteFrequencies;
-    for (size_t site=0; site<siteCount; ++site) {
+    for (intptr_t site=0; site<siteCount; ++site) {
         SiteSummary &s      = sites[site];
         bool alreadyCounted = false;
         totalFrequency     += s.frequency;
@@ -101,11 +101,11 @@ AlignmentSummary::~AlignmentSummary() {
     sequenceCount  = 0;
 }
 
-bool   AlignmentSummary::hasSequenceMatrix() const {
+bool AlignmentSummary::hasSequenceMatrix() const {
     return sequenceMatrix != nullptr;
 }
 
-size_t AlignmentSummary::getSequenceCount() const {
+int AlignmentSummary::getSequenceCount() const {
     return sequenceCount;
 }
 
@@ -134,12 +134,12 @@ const char* AlignmentSummary::getSequenceMatrix() const {
     return sequenceMatrix;
 }
 
-const char* AlignmentSummary::getSequence(size_t sequence_id) const {
+const char* AlignmentSummary::getSequence(int sequence_id) const {
     if (sequenceMatrix==nullptr) {
         return nullptr;
     }
     ASSERT(sequence_id < sequenceCount);
-    return sequenceMatrix + sequence_id * sequenceLength;
+    return sequenceMatrix + static_cast<intptr_t>(sequence_id) * sequenceLength;
 }
 
 size_t AlignmentSummary::getSequenceLength() const {
@@ -148,12 +148,12 @@ size_t AlignmentSummary::getSequenceLength() const {
 
 bool AlignmentSummary::constructSequenceMatrixNoisily(bool treatAllAmbiguousStatesAsUnknown
     , const char* taskName, const char* verb) {
-    progress_display progress(sequenceCount, taskName, verb, "sequence");
+    progress_display progress(static_cast<double>(sequenceCount), taskName, verb, "sequence");
     return constructSequenceMatrix(treatAllAmbiguousStatesAsUnknown, &progress);
 }
 
 size_t AlignmentSummary::getStateCount() const {
-    return maxState - minState + 1;
+    return static_cast<size_t>(maxState) - static_cast<size_t>(minState) + 1;
 }
 
 
@@ -165,17 +165,17 @@ bool AlignmentSummary::constructSequenceMatrix ( bool treatAllAmbiguousStatesAsU
         return false;
     }
     sequenceMatrix =  new char[ sequenceCount * sequenceLength ];
-    const int* posToSite = siteNumbers.data();
+    const intptr_t* posToSite = siteNumbers.data();
     if (treatAllAmbiguousStatesAsUnknown)
     {
         #ifdef _OPENMP
         #pragma omp parallel for
         #endif
-        for (size_t seq=0; seq<sequenceCount; ++seq) { //sequence
+        for (intptr_t seq=0; seq<sequenceCount; ++seq) { //sequence
             char *sequence = sequenceMatrix + seq * sequenceLength;
-            for (size_t seqPos = 0; seqPos < sequenceLength; ++seqPos) { //position in reduced sequence
+            for (intptr_t seqPos = 0; seqPos < sequenceLength; ++seqPos) { //position in reduced sequence
                 auto state = alignment->at(posToSite[seqPos])[seq] ;
-                if ( this->alignment->num_states <= state ) {
+                if ( this->alignment->num_states <= static_cast<int>(state) ) {
                     state = this->alignment->STATE_UNKNOWN;
                 }
                 sequence[seqPos] = static_cast<char> ( state );
@@ -191,9 +191,9 @@ bool AlignmentSummary::constructSequenceMatrix ( bool treatAllAmbiguousStatesAsU
         #ifdef _OPENMP
         #pragma omp parallel for
         #endif
-        for (size_t seq=0; seq<sequenceCount; ++seq) { //sequence
+        for (intptr_t seq=0; seq<sequenceCount; ++seq) { //sequence
             char *sequence = sequenceMatrix + seq * sequenceLength;
-            for (size_t seqPos = 0; seqPos < sequenceLength; ++seqPos) { //position in reduced sequence
+            for (intptr_t seqPos = 0; seqPos < sequenceLength; ++seqPos) { //position in reduced sequence
                 sequence[seqPos] = static_cast<char> ( alignment->at(posToSite[seqPos])[seq] );
                 //the state at the (seqPos)th non-constant site, in the (seq)th sequence
             }
